+++ src/dynamo/dynamo/inputplugins/packer.cpp
@@ -3719,7 +3719,7 @@ Vector
 CIPPacker::getRandVelVec()
 {
   //See http://mathworld.wolfram.com/SpherePointPicking.html
-  boost::normal_distribution<double> normdist(0.0, (1.0 / sqrt(NDIM)));
+  boost::normal_distribution<double> normdist(0.0, (1.0 / sqrt(double(NDIM))));
 
   boost::variate_generator<dynamo::baseRNG&, boost::normal_distribution<double> >
     normal_sampler(Sim->ranGenerator, normdist);
